+++ test/publishing_api_v2/lookup_test.rb
@@ -35,6 +35,33 @@ describe GdsApi::PublishingApiV2 do
 
       assert_equal "08f86d00-e95f-492f-af1d-470c5ba4752e", content_id
     end
+
+    it "returns the content_id of a draft document for a base_path" do
+      publishing_api
+        .given("there is a draft content item with base_path /foo")
+        .upon_receiving("a /lookup-by-base-path-request")
+        .with(
+          method: :post,
+          path: "/lookup-by-base-path",
+          body: {
+            base_paths: ["/foo"],
+            with_drafts: true,
+          },
+          headers: {
+            "Content-Type" => "application/json",
+          },
+        )
+        .will_respond_with(
+          status: 200,
+          body: {
+            "/foo" => "cbb460a7-60de-4a74-b5be-0b27c6d6af9b"
+          }
+        )
+
+      content_id = @api_client.lookup_content_id(base_path: "/foo", with_drafts: true)
+
+      assert_equal "cbb460a7-60de-4a74-b5be-0b27c6d6af9b", content_id
+    end
   end
 
   describe "#lookup_content_ids" do